requirements.txt
openai
tweepy
pandas
python-dotenv
duckduckgo_search
# tenacity
google-search-results
python-telegram-bot==13.11
